% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learner.ksvm-package}
\alias{mlr3learner.ksvm}
\alias{mlr3learner.ksvm-package}
\title{mlr3learner.ksvm: Wrapper for the kernlab::ksvm learner for use in mlr3}
\description{
mlr3 learner for 'ksvm' from the 'kernlab'
    package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr3learners/mlr3learners.ksvm}
}

}
\author{
\strong{Maintainer}: Marvin Böcker \email{marvin.boecker@tu-dortmund.de}

}
