% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.kernlab-package}
\alias{mlr3learners.kernlab}
\alias{mlr3learners.kernlab-package}
\title{mlr3learners.kernlab: Learners from kernlab package for mlr3}
\description{
Adds `ksvm()` from the kernlab package to mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr3learners/mlr3learners.kernlab}
}

}
\author{
\strong{Maintainer}: Marvin Böcker \email{marvin.boecker@tu-dortmund.de}

Authors:
\itemize{
  \item Marc Becker \email{marcbecker@posteo.de}
}

}
