% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrKSVM.R
\docType{data}
\name{LearnerRegrKSVM}
\alias{LearnerRegrKSVM}
\alias{mlr_learners_regr.ksvm}
\title{Regression Kernlab Support Vector Machine}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\usage{
LearnerRegrKSVM
}
\description{
A \link{LearnerRegr} for a regression support vector machine implemented in \code{\link[kernlab:ksvm]{kernlab::ksvm()}} in package \CRANpkg{kernlab}.
}
\references{
Alexandros Karatzoglou, Alex Smola, Kurt Hornik, Achim Zeileis (2004).
kernlab - An S4 Package for Kernel Methods in R. Journal of Statistical Software 11(9), 1-20.
\url{http://www.jstatsoft.org/v11/i09/}
}
\keyword{datasets}
